moved application user to identity folder
