+++ LibrarySystem/LibrarySystem/Infrastructure/Services/UserService.cs
@@ -1,5 +1,5 @@
 ﻿using LibrarySystem.Infrastructure.Interfaces;
-using LibrarySystem.Models;
+using LibrarySystem.Models.Identity;
 using LibrarySystem.Models.ViewModels;
 using LibrarySystem.Utils;
 using Microsoft.AspNetCore.Identity;
