+++ LibrarySystem/LibrarySystem/Models/Order.cs
@@ -1,4 +1,5 @@
 ﻿using LibrarySystem.Infrastructure.Enums;
+using LibrarySystem.Models.Identity;
 using Microsoft.EntityFrameworkCore;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
